% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmiss.R
\name{mmiss}
\alias{mmiss}
\title{Tabulating missing values/percentages}
\usage{
mmiss(df)
}
\description{
This function displays counts and percentages of missing values
for each variable as well as the entire dataset
}
\examples{
my_vec <- 1:5
my_vec2 <- c(1:5,NA)
v1 <- sample(my_vec, size = 10, replace = TRUE)
v2 <- sample(my_vec2, size = 10, replace = TRUE)
v3 <- sample(my_vec, size = 10, replace = TRUE)
v4 <- sample(my_vec2, size = 10, replace = TRUE)
df <- data.frame(v1,v2,v3,v4)
df
mmiss(df)
}
